updated osquery-go pkg, moved pkg/interactive to pkg/osquery/interactive, updates in response to comments
