+++ go.mod
@@ -49,7 +49,7 @@ require (
 	github.com/onsi/gomega v1.4.3 // indirect
 	github.com/opencontainers/go-digest v1.0.0-rc1 // indirect
 	github.com/opencontainers/image-spec v1.0.2 // indirect
-	github.com/osquery/osquery-go v0.0.0-20220317165851-954ac78f381f
+	github.com/osquery/osquery-go v0.0.0-20220706183148-4e1f83012b42
 	github.com/peterbourgon/ff/v3 v3.0.0
 	github.com/pkg/errors v0.8.1
 	github.com/prometheus/client_golang v0.9.2 // indirect
